requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
geographiclib==2.0
geopy==2.4.1
h3==3.7.7
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
pycparser==2.22
pyfiglet==1.0.2
Pygments==2.17.2
pynput==1.7.6
pyobjc-core==10.2
pyobjc-framework-ApplicationServices==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-CoreText==10.2
pyobjc-framework-Quartz==10.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rich==13.7.1
setuptools==69.2.0
six==1.16.0
timezonefinder==6.5.0
urllib3==2.2.1
